NavigationPlugin: rename to NavigationByMap
