+++ plugins/NavigationByMap/build.xml
@@ -1,3 +1,3 @@
-<project name="NavigationPlugin" default="jar">
+<project name="NavigationByMap" default="jar">
     <import file="../javapluginbuild.xml"/>
 </project>
